+++ src/components/series/post-list-on-series.tsx
@@ -1,8 +1,8 @@
 import Link from "next/link";
 import { type Post } from "@/schema/posts";
 import formatDate from "@/utils/format-date";
+import TagLink from "../tags/tag-link";
 import TagList from "../tags/tag-list";
-import TagListItem from "../tags/tag-list-item";
 
 type Props = {
   postsOnSeries: Post[];
@@ -34,7 +34,9 @@ export default function PostListOnSeries({ postsOnSeries }: Props) {
             {post.tags && post.tags.length > 0 ? (
               <TagList>
                 {post.tags.map((tag) => (
-                  <TagListItem key={tag} tag={tag} link={`/posts?tag=${tag}`} />
+                  <li key={tag}>
+                    <TagLink tag={tag} link={`/posts?tag=${tag}`} />
+                  </li>
                 ))}
               </TagList>
             ) : null}
